Cluster results by suggestion type.
This uses a ExpandableListView for results, with the first
group showing promoted suggestions, and the remaining groups
showing all suggestions by corpus type.

Bug: 3162608
Dependant on change: I579339ca208a83ecb9813175da7b8afbb3aa3c9f

+++ src/com/android/quicksearchbox/ui/SuggestionView.java
@@ -33,7 +33,7 @@ public interface SuggestionView {
      * @param adapter The adapter of the list which the view is appearing in
      * @param position The position of this view with the list.
      */
-    void bindAdapter(SuggestionsAdapter<?> adapter, int position);
+    void bindAdapter(SuggestionsAdapter<?> adapter, long position);
 
     /**
      * Call this with false to turn off the icon on the left. By default it is visible.
